refactoring: all funtions oved to different files dependeing it's funcitonalities.
